SlackConversationSetting に Card 追加
