+++ src/components/index.ts
@@ -4,4 +4,4 @@ export { User } from './User'
 export { PunchInForm } from './PunchInForm'
 export { SlackEmojiSettings } from './SlackEmojiSettings'
 export { LocalStorageError } from './LocalStorageError'
-export { SlackConversationSettings } from './SlackConversationSettings'
+export { SlackConversationSetting } from './SlackConversationSetting.tsx'
